update a file

+++ temp_c_to_f.py
@@ -1,4 +1,4 @@
 c = input('請輸入攝氏溫度 :')
-c = int(c)
+c = float(c)
 f = c * 9 / 5 + 32
 print ('攝氏 :', c ,' => ' '華氏 :', f)
